requirements.txt
requests==2.19.1
ortools==6.8.5452
Pillow==5.2.0